move toolbar sub components into new toolbar location
