+++ projects/media-viewer/src/lib/media-viewer/toolbar/left-pane/left-pane.component.html
