+++ projects/media-viewer/src/lib/media-viewer/toolbar/middle-pane/middle-pane.component.html
