+++ projects/media-viewer/src/lib/media-viewer/toolbar/right-pane/right-pane.component.html
