+++ projects/media-viewer/src/lib/media-viewer/toolbar/search-bar/search-bar.component.html
