+++ projects/media-viewer/src/lib/media-viewer/toolbar/side-bar/side-bar.component.ts
