+++ projects/media-viewer/src/lib/media-viewer/toolbar/styles/button-styles.scss
