+++ projects/media-viewer/src/lib/media-viewer/toolbar/styles/debugging.styles.scss
