+++ projects/media-viewer/src/lib/media-viewer/toolbar/styles/dialog-overlay.styles.scss
