+++ projects/media-viewer/src/lib/media-viewer/toolbar/styles/field-label.styles.scss
