+++ projects/media-viewer/src/lib/media-viewer/toolbar/styles/loading-bar.styles.scss
