+++ projects/media-viewer/src/lib/media-viewer/toolbar/styles/media-styles.scss
