+++ projects/media-viewer/src/lib/media-viewer/toolbar/styles/outline-attachment.styles.scss
