+++ projects/media-viewer/src/lib/media-viewer/toolbar/styles/search-bar.styles.scss
