+++ projects/media-viewer/src/lib/media-viewer/toolbar/styles/side-bar-styles.scss
