+++ projects/media-viewer/src/lib/media-viewer/toolbar/styles/sub-toolbar-styles.scss
