+++ projects/media-viewer/src/lib/media-viewer/toolbar/styles/text-selection.styles.scss
