+++ projects/media-viewer/src/lib/media-viewer/toolbar/styles/thumbnail.styles.scss
