+++ projects/media-viewer/src/lib/media-viewer/toolbar/styles/toolbar.component.scss
